tutorials/redundancy_removal/requirements.txt
numpy
paddlenlp
paddlepaddle-gpu >= 2.0.0
scikit-learn
tqdm
matplotlib
IPython
pre-commit